Added administrator requirements for delete operations.

+++ resources/user.py
@@ -71,9 +71,13 @@ class User(MethodView):
     def get(self, user_id):
         user = UserModel.query.get_or_404(user_id)
         return user
+    # @blp.alt_response(401, description="Unauthorized.")
     
     @jwt_required()
     def delete(self, user_id):
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         user = UserModel.query.get_or_404(user_id)
         db.session.delete(user)
         db.session.commit()
